+++ ui/helpers/constants.js
@@ -303,3 +303,20 @@ export const compareDefaultTimeRange = 86400 * 2;
 export const thBugSuggestionLimit = 20;
 
 export const thMaxPushFetchSize = 100;
+
+// we only want bug_ui and tree_ui to be used for UI validation
+// (for Intermittent Failures View) because if there is a valid
+// type used but its a non-existent repo or bug_id
+// we want to see that message from the api response
+export const prettyErrorMessages = {
+  startday: 'startday is required and must be in YYYY-MM-DD format.',
+  endday: 'endday is required and must be in YYYY-MM-DD format.',
+  bug_ui: 'bug is required and must be a valid integer.',
+  tree_ui:
+    'tree is required and must be a valid repository or repository group.',
+  default: 'Something went wrong.',
+  status503:
+    'There was a problem retrieving the data. Please try again in a minute.',
+};
+
+export const errorMessageClass = 'text-danger py-4 d-block text-center';
