+++ ui/helpers/errorMessage.js
@@ -1,3 +1,5 @@
+import { prettyErrorMessages } from './constants';
+
 /**
 This object contains a few constants and helper functions related to error
 message handling.
@@ -42,4 +44,26 @@ export const formatTaskclusterError = function formatTaskclusterError(e) {
   }
 
   return `${TC_ERROR_PREFIX}${errorMessage}`;
+};
+
+export const processErrorMessage = function processErrorMessage(
+  errorMessage,
+  status,
+) {
+  const messages = [];
+
+  if (status === 503) {
+    return [prettyErrorMessages.status503];
+  }
+
+  if (Object.keys(errorMessage).length > 0) {
+    for (const [key, value] of Object.entries(errorMessage)) {
+      if (prettyErrorMessages[key]) {
+        messages.push(prettyErrorMessages[key]);
+      } else {
+        messages.push(`${key}: ${value}`);
+      }
+    }
+  }
+  return messages || [prettyErrorMessages.default];
 };
