+++ ui/intermittent-failures/helpers.js
@@ -1,6 +1,6 @@
 import moment from 'moment';
 
-import { prettyErrorMessages } from './constants';
+import { prettyErrorMessages } from '../helpers/constants';
 
 // be sure to wrap date arg in a moment()
 export const ISODate = function formatISODate(date) {
@@ -90,28 +90,6 @@ export const sortData = function sortData(data, sortBy, desc) {
     return 0;
   });
   return data;
-};
-
-export const processErrorMessage = function processErrorMessage(
-  errorMessage,
-  status,
-) {
-  const messages = [];
-
-  if (status === 503) {
-    return [prettyErrorMessages.status503];
-  }
-
-  if (Object.keys(errorMessage).length > 0) {
-    for (const [key, value] of Object.entries(errorMessage)) {
-      if (prettyErrorMessages[key]) {
-        messages.push(prettyErrorMessages[key]);
-      } else {
-        messages.push(`${key}: ${value}`);
-      }
-    }
-  }
-  return messages || [prettyErrorMessages.default];
 };
 
 export const validateQueryParams = function validateQueryParams(
