+++ ui/partials/perf/comparechooserctrl.html
@@ -1,3 +1,1 @@
-<div>
-  <compare-selector-view />
-</div>
+<compare-selector-view />
